+++ ExampleUITests/ExampleUITests.swift
@@ -5,12 +5,11 @@
 //  Created by Krunoslav Zaher on 1/1/16.
 //  Copyright © 2016 kzaher. All rights reserved.
 //
-
 import XCTest
 import CoreLocation
 
 class ExampleUITests: XCTestCase {
-        
+
     override func setUp() {
         super.setUp()
 
@@ -21,7 +20,7 @@ class ExampleUITests: XCTestCase {
     override func tearDown() {
         super.tearDown()
     }
-    
+
     func testExample() {
         XCUIApplication().tables.element(boundBy: 0).cells.staticTexts["Randomize Example"].tap()