+++ dependencies.sh
@@ -1,0 +1,6 @@
+#set -e
+#if [[ ( ! -d "RxSwift/.git" ) && ( ! -d "Carthage" ) && ( ! -d 'Pods' ) ]]; then
+#    git submodule update --init --recursive --force
+#    cd RxSwift
+#    git reset origin/master --hard
+#fi